requirements-test.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0
jsonschema==4.20.0
pytest-xdist==3.3.1
pytest-timeout==2.1.0 